backend/lambdas/auth/requirements.txt
boto3>=1.28.0
PyJWT>=2.8.0